+++ tests/testthat/test-f7Picker.R
@@ -1,0 +1,3 @@
+test_that("picker works", {
+  expect_equal(2 * 2, 4)
+})